Fixed TwoLibertyHeuristic constructor AGAIN

+++ Orego/src/orego/heuristic/TwoLibertyHeuristic.java
@@ -18,7 +18,7 @@ public class TwoLibertyHeuristic extends Heuristic {
 
 	private int weight;
 
-	public TwoLibertyHeuristic(int weight2) {
+	public TwoLibertyHeuristic(double weight2) {
 		setWeight(weight2);
 		targets = new IntList(4);
 		weight = 1;
